requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt
